Randomized border colour, and sped up calculation time.
I've made the border color random, so it's easier to identify the various squares and circles from each other.  A better solution would be to cycle through a set of predetermined colours, but that might be overthinking it.

I've also sped up the time to calculate everything, by only checking every other pixel for a hotspot. This gets rid of the noticable delay when you click the key that draws everything. The only downside is that there's only a 25% chance of finding a 1x1 pixel hotspot.
I've made the starting location for calculation randomized so there's a chance it can pick up a 1x1 pixel hotspot with enough button presses. But this also comes with the cost of the shapes sometimes moving a little bit on repeated button presses.